DarwinG-Marketing/pyproject.toml
[project]
name = "darwing-marketing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.11.3",
    "python-dotenv>=1.1.1",
    "virtuals-tweepy>=0.1.6",
]
